requirements.txt
openai
pyyaml
pandas
wordcloud
matplotlib
statsmodels
openpyxl
trueskill
scipy
seaborn